Minor fixes

+++ Microcontroller/include/SPI_comm.h
@@ -5,6 +5,6 @@
 
 // SPI function declarations
 void spi_stress_test_10000_packets(void);
-void transmit_signalgenerator_data(uint16_t amp, uint8_t freq, uint8_t shape);
+void transmit_signalgenerator_data(uint8_t amp, uint8_t freq, uint8_t shape);
 
 #endif // SPI_H
